Refactor commands into separate custom ones
The idea here would be to keep all the common commands configuration in a commonCommands.conf file for example, then the commands.conf file would start with the line {commonCommands.conf}, and then would contain the configuration for the boss and the kick commands for example. Then you would have a permBossCommands.conf that would start exactly the same way ( {commonCommands.conf}), but the remaining lines for the configuration of the boss and the kick commands would be different. That way you would factorize as much configuration as possible.