fix(battlePresets): update regex to match the base64 format used by BAR
The existing regex uses `+` and `/` for characters 62 and 63, which is
"standard" base64, while the game uses `-` and `_` for those characters,
which is "URL-safe" base64.

The result of that mismatch is that many tweakdefs are impossible to
enter, as they include `-` or `_`. This is often possible to work
around (by adding extra spaces or dummy characters), but particularly
for more complicated scripts, this can be quite difficult.

This commit switches the filter to match "URL-safe" base64 so that all
properly encoded tweakdefs can be entered.