Prevent electing bosses for lobbies with more than one player.
This patch makes two changes:

First, logic is added to `preSpadsCommand()` which blocks commands
of the form "!callvote boss USERNAME" when:
- There is more than one player in the lobby
- AND, the caller has accessLevel < 100 (not a boss/mod/admin)

(Note that SPADS implicitly converts "!boss" to "!callvote boss"
when the caller is low-privileged.)

Second, spectators with accessLevel 100 (already a boss) are
permitted to use the "!boss" command. This makes it more convenient
to transfer boss privileges to another player.