Set an empty Teaser for private lobbies
Currently, sendTachyonBattleTeaser() returns early if the lobby
has a password set. With this change, an empty string is used
as the Teaser instead.

This change is intended to make the function more robust; if a
Teaser happens to be set when it shouldn't be, then the function
now has a way to correct that.