Update Teaser when the first player joins a lobby
BarManager removes the Teaser when the last player leaves a lobby.

But, currently it does not have a cooresponding check for whether a
Teaser should be set when a first player joins the lobby. This
patch adds such a check; this makes it so that the Teaser will be
updated quickly (before a configuration change is made).