Make Teaser "Boss" entry consistent between presets
Currently, the placement and formating of the "Boss" portion of
Teasers is inconsistent between the Custom preset and all other
presets: it moves between first/last entry and has its |-separator
disappear when the Custom preset is selected.

This patch moves the "Boss" logic so that it is no longer dependent
on what preset is selected. Now, the "Boss" entry is always first,
and always has a proper |-separator.