Make Teasers less verbose (#145)
* Merge Integration into Master (Fix Team Ratings) (#140)

* Fix ratings not updating in Chobby

* Check for change when preset applied

* Minor spelling mistake

Minor spelling mistake

* Updates after Beherith feedback

---------

Co-authored-by: Joshua Augustinus <jauggy@gmail.com>

* Make Teaser "Boss" entry consistent between presets

Currently, the placement and formating of the "Boss" portion of
Teasers is inconsistent between the Custom preset and all other
presets: it moves between first/last entry and has its |-separator
disappear when the Custom preset is selected.

This patch moves the "Boss" logic so that it is no longer dependent
on what preset is selected. Now, the "Boss" entry is always first,
and always has a proper |-separator.

* Fix check for outdated "ChickensAI" in sendTachyonBattleTeaser()

Replaced with the modern name, "RaptorsAI".

* Choose only a single AI type to include in Teasers.

Currently, all unique AI types are listed in the Teaser. This
quickly becomes very long, and is not generally useful.

Instead, just pick a single AI type among those in the lobby, and
leave the rest out of the Teaser.

* Shorten Teaser descriptions for team size/number settings.

Before, the Teaser might include "4 vs 4 vs 4 vs 4". With this
change, that configuration would instead look like "4v4v4v4".

The Team preset is also now shown as "Team"; before, it would
appear as "Teams".

* Include either "Boss:" or "Preset+Config" in Teasers, not both.

---------

Co-authored-by: Beherith <mysterme@gmail.com>
Co-authored-by: Joshua Augustinus <jauggy@gmail.com>