Completely remove "Boss:" from Teasers, and hide preset when sane.
Previously, a "Boss:" Teaser would conditionally be used. With this
patch, the "Boss:" entry is never included in the Teaser.

Additionally, when the "Team" and "Custom" presets are in use (and
there are no bots), the name of the preset is not displayed. For
example:

Before change: ` | Team 8v8`
After change:  ` | 8v8`

Before change: ` | Custom 4v4v4v4`
After change: ` | 4v4v4v4`

This does not apply when bots are configured. In such cases, the
preset name is still included (same behavior as before). For
example:

- ` | Team vs Scavengers`
- ` | Custom vs Raptors`
- ` | Coop vs AI`