Integration merging into Main (#147)
* Fix ratings not updating in Chobby

* Check for change when preset applied

* Minor spelling mistake

Minor spelling mistake

* Updates after Beherith feedback

* Make Teasers less verbose (#145)

* Merge Integration into Master (Fix Team Ratings) (#140)

* Fix ratings not updating in Chobby

* Check for change when preset applied

* Minor spelling mistake

Minor spelling mistake

* Updates after Beherith feedback

---------

Co-authored-by: Joshua Augustinus <jauggy@gmail.com>

* Make Teaser "Boss" entry consistent between presets

Currently, the placement and formating of the "Boss" portion of
Teasers is inconsistent between the Custom preset and all other
presets: it moves between first/last entry and has its |-separator
disappear when the Custom preset is selected.

This patch moves the "Boss" logic so that it is no longer dependent
on what preset is selected. Now, the "Boss" entry is always first,
and always has a proper |-separator.

* Fix check for outdated "ChickensAI" in sendTachyonBattleTeaser()

Replaced with the modern name, "RaptorsAI".

* Choose only a single AI type to include in Teasers.

Currently, all unique AI types are listed in the Teaser. This
quickly becomes very long, and is not generally useful.

Instead, just pick a single AI type among those in the lobby, and
leave the rest out of the Teaser.

* Shorten Teaser descriptions for team size/number settings.

Before, the Teaser might include "4 vs 4 vs 4 vs 4". With this
change, that configuration would instead look like "4v4v4v4".

The Team preset is also now shown as "Team"; before, it would
appear as "Teams".

* Include either "Boss:" or "Preset+Config" in Teasers, not both.

---------

Co-authored-by: Beherith <mysterme@gmail.com>
Co-authored-by: Joshua Augustinus <jauggy@gmail.com>

* Update teaser on first player join (#144)

* Merge Integration into Master (Fix Team Ratings) (#140)

* Fix ratings not updating in Chobby

* Check for change when preset applied

* Minor spelling mistake

Minor spelling mistake

* Updates after Beherith feedback

---------

Co-authored-by: Joshua Augustinus <jauggy@gmail.com>

* Update Teaser when the first player joins a lobby

BarManager removes the Teaser when the last player leaves a lobby.

But, currently it does not have a cooresponding check for whether a
Teaser should be set when a first player joins the lobby. This
patch adds such a check; this makes it so that the Teaser will be
updated quickly (before a configuration change is made).

---------

Co-authored-by: Beherith <mysterme@gmail.com>
Co-authored-by: Joshua Augustinus <jauggy@gmail.com>

* Prevent electing bosses for lobbies with more than one player. (#138)

* Prevent electing bosses for lobbies with more than one player.

This patch makes two changes:

First, logic is added to `preSpadsCommand()` which blocks commands
of the form "!callvote boss USERNAME" when:
- There is more than one player in the lobby
- AND, the caller has accessLevel < 100 (not a boss/mod/admin)

(Note that SPADS implicitly converts "!boss" to "!callvote boss"
when the caller is low-privileged.)

Second, spectators with accessLevel 100 (already a boss) are
permitted to use the "!boss" command. This makes it more convenient
to transfer boss privileges to another player.

* Allow spectating bosses to use '!boss ...' with non-custom presets.

* Set empty teaser for private lobbies (#143)

* Merge Integration into Master (Fix Team Ratings) (#140)

* Fix ratings not updating in Chobby

* Check for change when preset applied

* Minor spelling mistake

Minor spelling mistake

* Updates after Beherith feedback

---------

Co-authored-by: Joshua Augustinus <jauggy@gmail.com>

* Set an empty Teaser for private lobbies

Currently, sendTachyonBattleTeaser() returns early if the lobby
has a password set. With this change, an empty string is used
as the Teaser instead.

This change is intended to make the function more robust; if a
Teaser happens to be set when it shouldn't be, then the function
now has a way to correct that.

---------

Co-authored-by: Beherith <mysterme@gmail.com>
Co-authored-by: Joshua Augustinus <jauggy@gmail.com>

* Completely remove "Boss:" from Teasers, and hide preset when sane. (#146)

Previously, a "Boss:" Teaser would conditionally be used. With this
patch, the "Boss:" entry is never included in the Teaser.

Additionally, when the "Team" and "Custom" presets are in use (and
there are no bots), the name of the preset is not displayed. For
example:

Before change: ` | Team 8v8`
After change:  ` | 8v8`

Before change: ` | Custom 4v4v4v4`
After change: ` | 4v4v4v4`

This does not apply when bots are configured. In such cases, the
preset name is still included (same behavior as before). For
example:

- ` | Team vs Scavengers`
- ` | Custom vs Raptors`
- ` | Coop vs AI`

---------

Co-authored-by: Joshua Augustinus <jauggy@gmail.com>
Co-authored-by: DeviousNull <Fr0z3n.0n3@gmail.com>