Send all bosses at once in BarManager "boss" updates.
Currently, BarManager's "boss" state updates consist of either an
empty string (to remove all bosses) or a single username (to add an
additional boss).

This causes issues when a lobby has multiple bosses. For example,
a newly-joining player will only see the crown for a single boss
of the lobby, even if that lobby has multiple bosses at the time.

This patch seeks to resolve the root of this issue, by including
all bosses in BarManager's "boss" state updates (separated by
commas).

*A change is required on the client's side to support the new
update format.* If an unmodified client joins a lobby with this
change, then that client will only see the crown for a single boss
at any single point in time (regardless of how many bosses are
present). Since multi-boss lobbies are relatively rare, and this
is only a cosmetic issue, there should not be any significant
impact to players before the new code is in use for both
BYAR-Chobby and BarManager.