Log LUAMSG events for pauses/unpauses via a fake chat message.
This patch adds a new case to the existing GAME_LUAMSG handler, for
when the received LUAMSG is in the format `p@u$3:%s` (where %s is
true/false).

When that LUAMSG is received, a fake chat message is sent to
Teiserver's `AutohostMonitor`, where it will be logged as part of
the match's chat log. This fake chat message is not displayed to
players/spectators.

The fake chat message includes the username of the player
who paused/unpaused the game, and what the current pause/unpause
state is.