Send "mutes" data in Chobby updates, containing muted usernames.
This patch allows BarManager to query the `InGameMutes` plugin
(if it is loaded) for the muted/unmuted status of each player in
the lobby.

The resulting list of muted users is sent to Chobby with all other
BarManager state, so that it can be displayed to users.

The following events trigger a re-query of `InGameMutes` (and
potentially the sending of an update):
- The `!mute` command is run successfully
- The `!unmute` command is run successfully
- A player joins the lobby
- A player leaves the lobby