Merge Integration -> Main (#169)
* Add some missing calls to 'removeSpadsCommandHandler'.

* Send all bosses at once in BarManager "boss" updates.

Currently, BarManager's "boss" state updates consist of either an
empty string (to remove all bosses) or a single username (to add an
additional boss).

This causes issues when a lobby has multiple bosses. For example,
a newly-joining player will only see the crown for a single boss
of the lobby, even if that lobby has multiple bosses at the time.

This patch seeks to resolve the root of this issue, by including
all bosses in BarManager's "boss" state updates (separated by
commas).

*A change is required on the client's side to support the new
update format.* If an unmodified client joins a lobby with this
change, then that client will only see the crown for a single boss
at any single point in time (regardless of how many bosses are
present). Since multi-boss lobbies are relatively rare, and this
is only a cosmetic issue, there should not be any significant
impact to players before the new code is in use for both
BYAR-Chobby and BarManager.

* Log LUAMSG events for pauses/unpauses via a fake chat message.

This patch adds a new case to the existing GAME_LUAMSG handler, for
when the received LUAMSG is in the format `p@u$3:%s` (where %s is
true/false).

When that LUAMSG is received, a fake chat message is sent to
Teiserver's `AutohostMonitor`, where it will be logged as part of
the match's chat log. This fake chat message is not displayed to
players/spectators.

The fake chat message includes the username of the player
who paused/unpaused the game, and what the current pause/unpause
state is.

* Ring potential voters when a vote begins.

This uses the same code as the `!ring` command, so ring-spam
protection applies if there are rapid votes.

* Reduce the default `voteTime` from 45 to 30 seconds.

* Make handlers for InGameMute plugin commands return "1" on success.

This allows other plugins to determine whether `!mute`, `!unmute`,
and `!mutes` succeed or fail.

* Send "mutes" data in Chobby updates, containing muted usernames.

This patch allows BarManager to query the `InGameMutes` plugin
(if it is loaded) for the muted/unmuted status of each player in
the lobby.

The resulting list of muted users is sent to Chobby with all other
BarManager state, so that it can be displayed to users.

The following events trigger a re-query of `InGameMutes` (and
potentially the sending of an update):
- The `!mute` command is run successfully
- The `!unmute` command is run successfully
- A player joins the lobby
- A player leaves the lobby

* Remove leftover DEBUG logging statement.

* Make direct calls to SPADS's Perl functions more robust.

Directly calling a non-existant SPADS function would raise an
exception in Perl code, which cannot be caught from within
Python code.

Therefore, in case SPADS renames one of its handler functions in
the future, we should always check that a Perl function actually
exists before calling it.

This patch adds a helper function `callPerlFunction(name, *args)`,
which only calls a SPADS function if it exists, and converts all
existing SPADS function calls to use the helper function.

* Allow force spec to deal with afk boss

---------

Co-authored-by: Joshua Augustinus <jauggy@gmail.com>