Replace `playersInMyBattle` with relevant call to the SPADS API.
The global variable `playersInMyBattle` was a dict which mapped
usernames to the initial battleStatus for that user, when they
joined the battle. However:
- The "battleStatus" value was never actually used (the whole dict
 was really just to get a count of users in the battle)
- The logic to remove entries from the dict was buggy; if a user
 joined with battleStatus "0" (rare, but possible!), then they
 would never be removed.

Since it is easy to get a count of users from the SPADS API, this
patch removes `playersInMyBattle`, and converts all previous uses
of that variable with a new helper function
`getNumUsersInMyBattle()`. This avoids duplicating the same data
between SPADS and BarManager, and removes one source of potential
out-of-sync bugs between the two.