Merge Integration->Main (#176)
* Add `!unboss` command.

Add a new `!unboss` command, which:
- Allows players to remove an individual boss in multi-boss lobbies
 (in the form `!unboss NameOfUser`)
- Allows players to remove all bosses at once
 (in the form `!unboss *`, similar to the current behavior of
 `!boss`)
- Has an instructional error message when run without parameters
 (to avoid the pitfall for `!boss`, where users often thought that
 running that command without parameters would only unboss
 themself and not other bosses)

This patch specifically *does not* change the behavior of the
`!boss` command (without parameters). Once Chobby has been updated
to send `!unboss USERNAME` instead of `!boss`, a new patch can be
applied to disable `!boss` (without parameters) with an
instructional message.

* Use purpose-built SPADS response functions in command handlers.

* Add missing `fix_string` call in `hUnboss` function.

* Remove unused `myCommand` command handler.

* Remove unused `splitbattle` command handler.

* Use more-standard perl syntax in `unboss` handler.

* Remove outdated `!meme` options to match Teiserver.

This patch brings BarManager in line with https://github.com/beyond-all-reason/teiserver/pull/476

* Replace `playersInMyBattle` with relevant call to the SPADS API.

The global variable `playersInMyBattle` was a dict which mapped
usernames to the initial battleStatus for that user, when they
joined the battle. However:
- The "battleStatus" value was never actually used (the whole dict
 was really just to get a count of users in the battle)
- The logic to remove entries from the dict was buggy; if a user
 joined with battleStatus "0" (rare, but possible!), then they
 would never be removed.

Since it is easy to get a count of users from the SPADS API, this
patch removes `playersInMyBattle`, and converts all previous uses
of that variable with a new helper function
`getNumUsersInMyBattle()`. This avoids duplicating the same data
between SPADS and BarManager, and removes one source of potential
out-of-sync bugs between the two.