Fix missing `!help` entries for several BarManager commands.
This patch allows the `!help` command to show documentation for
the following commands:
- `!setAllAiBonus`
- `!unboss`
- `!setratinglevels`

The help entries for the first two of these was missing because the
permissions configuration was placed in the global `commands.conf`
file, instead of the plugin-specific `BarManagerCmd.conf` file.

The help entry for the last command was missing because of a
copy-paste error (`!resetratinglevels` had two help entries, while
`!setratinglevels` had none).