Don't attempt to send battle updates when the battle is closed
Currently, `sendTachyonBattle()` is called as a timer callback, to
provide a form of rate-limiting. The timer is only started when a
battle is open, but it is possible that the lobby will disconnect
before `sendTachyonBattle()` runs.

This patch makes `sendTachyonBattle()` return early if the lobby
is no longer connected at the time that it runs.