Clarify an error message in various lobby command handlers
Currently, the following lobby command handlers log an error if
they are called for a user that is not currently known:
- hLEFT_pre
- hLEFTBATTLE_pre
- hREMOVEUSER_pre
- hCLIENTSTATUS_pre
- hJOINEDBATTLE_pre

When it is logged, the current error message states "hXXX_pre
cannot be exectuted <...>". This message does not clearly explain
what went wrong or what else may be impacted as a result of the
error.

This patch changes the error message to state "hXXX_pre was called
for an unknown user, skipping <...>". This makes it much more
clear what went wrong (missing a user), and what may be impacted
(anything involving that specific user).