Add `selfFileHash`, and print it in `!barmanagerprintstate`
Currently, there is not a reliable way to determine whether a host
is running the latest BarManager code. This is particularly
annoying when making changes to the Integration server, as for
subtle changes it is not easy to tell whether the latest code has
been fetched yet.

This patch adds a new `selfFileHash` variable (set on plugin
init), which calculates a 6-character hash of the currently-running
python file. That variable is also added to the output of
`!barmanagerprintstate`, so that it is easy to look up.