Make `!joinas` require approval exclusively from its target player.
Currently, if an arbitraty player runs the command `!joinas USER`,
a vote is called among all players in the lobby. As a result,
even if the target player does not want someone to be able to
control their units, the rest of the lobby can overrule them.

This patch changes how eligible voters are chosen for the `!joinas`
command as follows:
- If someone runs `!joinas spec` or `!joinas SELF_USERNAME`, then
 the previous behavior is maintained.
- If someone runs `!joinas OTHER_USERNAME`, and OTHER_USERNAME is
 not one of the eligible voters, then the vote is canceled.
- Otherwise, all voters *except* OTHER_USERNAME are made
 ineligible, and the vote is allowed to progress.

The practical result of the above, is that only the target of a
`!joinas` command is allowed to decide whether someone can join on
their team.