Make `!joinas` require approval from its target player.
Currently, if an arbitrary player runs the command `!joinas USER`,
a vote is called among all players in the lobby. As a result,
even if the target player does not want someone to be able to
control their units, the rest of the lobby can overrule them.

This patch makes the following changes, active during any ongoing
`!joinas USERNAME` vote:
- When a non-target submits a vote, the vote state is checked for
 whether USERNAME has voted already. If so, the non-target's vote
 is counted as normal; if not, the non-target's vote is deferred.
- If USERNAME submits a "no" vote, then the ongoing vote is
 immediately ended (as if by `!endVote`) and all deferred votes are
 discarded.
- When USERNAME submits a "yes" or "blank" vote, all deferred votes
 are counted.

This has the practical effect of requiring approval from the
`!joinas` target, in order for the vote to succeed; the vote will
fail if the target does not explicitly vote "yes" or "blank".