Add debugcommands to the coop preset
This change updates the coop battle preset to allow debugcommands. Currently, if a player wants to setup a coop vs ai lobby with terrain modifiers, it must be done with the team or custom preset. This causes issues because lobby balance must be turned off to keep the ai on one team. When balance is off, players can get assigned the same Id as an AI. The current solution is to wait for the lobby to fill up with the exact number of players, lock the lobby, and add re-add the ai.